refactor(ui): Remove unnecessary layout directory
The layout directory wasn't used for anything, so removing it simplifies
the paths.
