+++ ui/src/components/header.tsx
@@ -90,7 +90,7 @@ export const Header = () => {
   const runMatch = matches.find(
     (match) =>
       match.routeId ===
-      '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_layout/runs/$runIndex'
+      '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex'
   );
 
   return (
