+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex/config/-components/reporter-job-details.tsx
