+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex/vulnerabilities/index.tsx
@@ -17,12 +17,11 @@
  * License-Filename: LICENSE
  */
 
-import { createFileRoute } from '@tanstack/react-router';
+import { createFileRoute, Link } from '@tanstack/react-router';
 import {
   createColumnHelper,
   getCoreRowModel,
   getExpandedRowModel,
-  getFilteredRowModel,
   getPaginationRowModel,
   getSortedRowModel,
   Row,
@@ -31,17 +30,14 @@ import {
 import { ChevronDown, ChevronUp } from 'lucide-react';
 import { useMemo } from 'react';
 
+import { useVulnerabilitiesServiceGetVulnerabilitiesByRunId } from '@/api/queries';
 import { prefetchUseRepositoriesServiceGetOrtRunByIndex } from '@/api/queries/prefetch';
-import {
-  useRepositoriesServiceGetOrtRunByIndexSuspense,
-  useRuleViolationsServiceGetRuleViolationsByRunIdSuspense,
-} from '@/api/queries/suspense';
-import { RuleViolation } from '@/api/requests';
+import { useRepositoriesServiceGetOrtRunByIndexSuspense } from '@/api/queries/suspense';
+import { VulnerabilityWithIdentifier } from '@/api/requests';
 import { DataTable } from '@/components/data-table/data-table';
-import { DataTableToolbar } from '@/components/data-table/data-table-toolbar';
-import { FilterMultiSelect } from '@/components/data-table/filter-multi-select';
 import { LoadingIndicator } from '@/components/loading-indicator';
 import { MarkdownRenderer } from '@/components/markdown-renderer';
+import { ToastError } from '@/components/toast-error';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import {
@@ -51,26 +47,34 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
-import { getRuleViolationSeverityBackgroundColor } from '@/helpers/get-status-class';
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table';
+import { calcOverallVulnerability } from '@/helpers/calc-overall-vulnerability';
+import { getVulnerabilityRatingBackgroundColor } from '@/helpers/get-status-class';
 import { identifierToString } from '@/helpers/identifier-to-string';
-import { compareSeverity } from '@/helpers/sorting-functions';
+import { compareVulnerabilityRating } from '@/helpers/sorting-functions';
+import { toast } from '@/lib/toast';
 import {
   paginationSearchParameterSchema,
-  severitySchema,
-  severitySearchParameterSchema,
   sortingSearchParameterSchema,
 } from '@/schemas';
 
 const defaultPageSize = 10;
 
-const columnHelper = createColumnHelper<RuleViolation>();
+const columnHelper = createColumnHelper<VulnerabilityWithIdentifier>();
 
 const columns = [
   columnHelper.display({
     id: 'moreInfo',
     header: 'Details',
     size: 50,
-    cell: function CellComponent({ row }) {
+    cell: ({ row }) => {
       return row.getCanExpand() ? (
         <Button
           variant='outline'
@@ -92,69 +96,131 @@ const columns = [
     },
     enableSorting: false,
   }),
-  columnHelper.accessor('severity', {
-    header: 'Severity',
-    cell: ({ row }) => {
-      return (
-        <Badge
-          className={`${getRuleViolationSeverityBackgroundColor(row.original.severity)}`}
-        >
-          {row.original.severity}
-        </Badge>
-      );
-    },
-    filterFn: (row, _columnId, filterValue): boolean => {
-      return filterValue.includes(row.original.severity);
-    },
-    sortingFn: (rowA, rowB) => {
-      return compareSeverity(rowA.original.severity, rowB.original.severity);
-    },
-  }),
   columnHelper.accessor(
-    (ruleViolation) => {
-      return identifierToString(ruleViolation.packageId);
+    (vuln) => {
+      const ratings = vuln.vulnerability.references.map(
+        (reference) => reference.severity
+      );
+      return calcOverallVulnerability(ratings);
     },
     {
-      header: 'Package',
-      cell: ({ getValue }) => {
-        return <div className='font-semibold'>{getValue()}</div>;
+      id: 'rating',
+      header: 'Rating',
+      cell: ({ row }) => {
+        return (
+          <Badge
+            className={`${getVulnerabilityRatingBackgroundColor(row.getValue('rating'))}`}
+          >
+            {row.getValue('rating')}
+          </Badge>
+        );
+      },
+      sortingFn: (rowA, rowB) => {
+        return compareVulnerabilityRating(
+          rowA.getValue('rating'),
+          rowB.getValue('rating')
+        );
       },
     }
   ),
-  columnHelper.accessor('rule', {
-    header: 'Rule',
+  columnHelper.accessor(
+    (vuln) => {
+      return identifierToString(vuln.identifier);
+    },
+    {
+      id: 'package',
+      header: 'Package',
+      cell: ({ row }) => {
+        return <div className='font-semibold'>{row.getValue('package')}</div>;
+      },
+    }
+  ),
+  columnHelper.accessor('vulnerability.externalId', {
+    id: 'externalId',
+    header: 'External ID',
     cell: ({ row }) => (
       <Badge className='whitespace-nowrap bg-blue-300'>
-        {row.original.rule}
+        {row.getValue('externalId')}
       </Badge>
     ),
   }),
+  columnHelper.accessor(
+    (row) => {
+      return row.vulnerability.summary;
+    },
+    {
+      id: 'summary',
+      header: 'Summary',
+      cell: ({ row }) => {
+        return (
+          <div className='italic text-muted-foreground'>
+            {row.getValue('summary')}
+          </div>
+        );
+      },
+      enableSorting: false,
+    }
+  ),
 ];
 
-const renderSubComponent = ({ row }: { row: Row<RuleViolation> }) => {
-  const ruleViolation = row.original;
+const renderSubComponent = ({
+  row,
+}: {
+  row: Row<VulnerabilityWithIdentifier>;
+}) => {
+  const vulnerability = row.original.vulnerability;
 
   return (
     <div className='flex flex-col gap-4'>
-      <div>{ruleViolation.message}</div>
-      <div className='grid grid-cols-8 gap-2'>
-        <div className='col-span-2 font-semibold'>License:</div>
-        <div className='col-span-6'>{ruleViolation.license}</div>
-        <div className='col-span-2 font-semibold'>License source:</div>
-        <div className='col-span-6'>{ruleViolation.licenseSource}</div>
-        <div className='col-span-2 font-semibold'>How to fix:</div>
+      <div className='text-lg font-semibold'>Description</div>
+      <MarkdownRenderer
+        markdown={vulnerability.description || 'No description.'}
+      />
+      <div className='mt-2 text-lg font-semibold'>
+        Links to vulnerability references
       </div>
-      <MarkdownRenderer markdown={ruleViolation.howToFix} />
+      <Table>
+        <TableHeader>
+          <TableRow>
+            <TableHead>Severity</TableHead>
+            <TableHead>Scoring system</TableHead>
+            <TableHead>Score</TableHead>
+            <TableHead>Vector</TableHead>
+            <TableHead>Link</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {vulnerability.references.map((reference, index) => (
+            <TableRow key={index}>
+              <TableCell>{reference.severity || '-'}</TableCell>
+              <TableCell>{reference.scoringSystem || '-'}</TableCell>
+              <TableCell>{reference.score || '-'}</TableCell>
+              <TableCell>{reference.vector || '-'}</TableCell>
+              <TableCell>
+                {
+                  <Link
+                    className='break-all font-semibold text-blue-400 hover:underline'
+                    to={reference.url}
+                    target='_blank'
+                  >
+                    {reference.url}
+                  </Link>
+                }
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
     </div>
   );
 };
 
-const RuleViolationsComponent = () => {
+const VulnerabilitiesComponent = () => {
   const params = Route.useParams();
   const search = Route.useSearch();
-  const navigate = Route.useNavigate();
 
-  // Memoize the search parameters to prevent unnecessary re-rendering
+  // All of these need to be memoized to prevent unnecessary re-renders
+  // and (at least for Firefox) the browser freezing up.
 
   const pageIndex = useMemo(
     () => (search.page ? search.page - 1 : 0),
@@ -164,16 +230,6 @@ const RuleViolationsComponent = () => {
   const pageSize = useMemo(
     () => (search.pageSize ? search.pageSize : defaultPageSize),
     [search.pageSize]
-  );
-
-  const severity = useMemo(
-    () => (search.severity ? search.severity : undefined),
-    [search.severity]
-  );
-
-  const columnFilters = useMemo(
-    () => (severity ? [{ id: 'severity', value: severity }] : []),
-    [severity]
   );
 
   const sortBy = useMemo(() => {
@@ -190,73 +246,67 @@ const RuleViolationsComponent = () => {
     ortRunIndex: Number.parseInt(params.runIndex),
   });
 
-  const { data: ruleViolations } =
-    useRuleViolationsServiceGetRuleViolationsByRunIdSuspense({
-      runId: ortRun.id,
-      // Fetch all data at once, as we need to do both grouping and
-      // pagination in front-end for consistency in data handling.
-      limit: 100000,
-    });
+  const {
+    data: vulnerabilities,
+    isPending,
+    isError,
+    error,
+  } = useVulnerabilitiesServiceGetVulnerabilitiesByRunId({
+    runId: ortRun.id,
+    limit: 100000,
+  });
 
   const table = useReactTable({
-    data: ruleViolations?.data || [],
+    data: vulnerabilities?.data || [],
     columns,
     state: {
       pagination: {
         pageIndex,
         pageSize,
       },
-      columnFilters,
       sorting: sortBy,
     },
     getCoreRowModel: getCoreRowModel(),
     getExpandedRowModel: getExpandedRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
     getPaginationRowModel: getPaginationRowModel(),
     getSortedRowModel: getSortedRowModel(),
     getRowCanExpand: () => true,
+    enableMultiSort: false,
   });
+
+  if (isPending) {
+    return <LoadingIndicator />;
+  }
+
+  if (isError) {
+    toast.error('Unable to load data', {
+      description: <ToastError error={error} />,
+      duration: Infinity,
+      cancel: {
+        label: 'Dismiss',
+        onClick: () => {},
+      },
+    });
+    return;
+  }
 
   return (
     <Card className='h-fit'>
       <CardHeader>
-        <CardTitle>Rule violations (ORT run global ID: {ortRun.id})</CardTitle>
+        <CardTitle>Vulnerabilities (ORT run global ID: {ortRun.id})</CardTitle>
         <CardDescription>
-          These are the rule violations found from the ORT run. Please note that
-          the status may change over time, as your project dependencies change.
-          Therefore, your project should be scanned for rule violations
-          regularly.
+          These are the vulnerabilities found currently in the project. Please
+          note that the vulnerability status may change over time, as your
+          project dependencies change. Therefore, your project should be scanned
+          for vulnerabilities regularly.
+        </CardDescription>
+        <CardDescription>
+          By clicking on "References" you can see more information about the
+          vulnerability. The overall severity rating is calculated based on the
+          highest severity rating found in the references.
         </CardDescription>
       </CardHeader>
       <CardContent>
-        <DataTableToolbar
-          filters={
-            <FilterMultiSelect
-              title='Rule Violation Severity'
-              options={severitySchema.options.map((severity) => ({
-                label: severity,
-                value: severity,
-              }))}
-              selected={severity || []}
-              setSelected={(severities) => {
-                navigate({
-                  search: {
-                    ...search,
-                    page: 1,
-                    severity: severities.length === 0 ? undefined : severities,
-                  },
-                });
-              }}
-            />
-          }
-          resetFilters={() => {
-            navigate({
-              search: { ...search, page: 1, severity: undefined },
-            });
-          }}
-          resetBtnVisible={severity !== undefined}
-          className='mb-2'
-        />
         <DataTable
           table={table}
           renderSubComponent={renderSubComponent}
@@ -302,17 +352,17 @@ const RuleViolationsComponent = () => {
 };
 
 export const Route = createFileRoute(
-  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_layout/runs/$runIndex/rule-violations/'
+  '/_layout/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex/vulnerabilities/'
 )({
-  validateSearch: paginationSearchParameterSchema
-    .merge(severitySearchParameterSchema)
-    .merge(sortingSearchParameterSchema),
+  validateSearch: paginationSearchParameterSchema.merge(
+    sortingSearchParameterSchema
+  ),
   loader: async ({ context, params }) => {
     await prefetchUseRepositoriesServiceGetOrtRunByIndex(context.queryClient, {
       repositoryId: Number.parseInt(params.repoId),
       ortRunIndex: Number.parseInt(params.runIndex),
     });
   },
-  component: RuleViolationsComponent,
+  component: VulnerabilitiesComponent,
   pendingComponent: LoadingIndicator,
 });
